Updated and Added Schemas to add 2022 Player Data.

+++ backend/src/main/java/au/superdraftfantasy/api/seasonSummary/SeasonSummaryEntity.java
@@ -103,6 +103,6 @@ public class SeasonSummaryEntity {
 
     private Integer hardnessRating;
 
-    private Integer freesRatio;
+    private double freesRatio;
 
 }
